In case we want to perform conflict detection in the future
The problem seems to be that of handling the document session.
When a ConcurrencyException is thrown it means the document
session is invalid and we need a new one. So for this to happen
it seems that the event store needs to manage the document session.
That complicates things a bit, obviously. I'll leave it be for now.